changed to float phi

+++ src/Enemy.cpp
@@ -5,7 +5,7 @@ void Enemy::setSpawn() {
     std::uniform_real_distribution<> randomAngle(0.0, 2.0 * M_PI);
     std::uniform_real_distribution<> randomDistance(400.0, 600.0);
 
-    double phi = 0.0;
+    float phi = 0.0;
     do {
         phi = static_cast<float>(randomAngle(gen));
     //! Prevent unavoidable hits
